refactor(serve): use dynamic import for subcommand
