+++ packages/opencode/src/index.ts
@@ -14,7 +14,7 @@ import { Installation } from "./installation"
 import { InstallationVersion } from "@opencode-ai/core/installation/version"
 import { NamedError } from "@opencode-ai/core/util/error"
 import { FormatError } from "./cli/error"
-import { ServeCommand } from "./cli/cmd/serve"
+import { ServeCommand } from "./cli/cmd/serve/command"
 import { Filesystem } from "@/util/filesystem"
 import { DebugCommand } from "./cli/cmd/debug/command"
 import { StatsCommand } from "./cli/cmd/stats"
